slimming down pnpm lockfile parsing
