+++ package.json
@@ -17,8 +17,6 @@
     "test": "jest"
   },
   "dependencies": {
-    "@pnpm/lockfile-file": "^3.0.7",
-    "@pnpm/logger": "^3.2.2",
     "@yarnpkg/lockfile": "^1.1.0",
     "find-up": "^4.1.0",
     "find-yarn-workspace-root": "^1.2.1",
