+++ src/__tests__/lockfile.test.ts
@@ -1,4 +1,3 @@
-import * as path from "path";
 import { setupFixture } from "../helpers/setupFixture";
 import { parseLockFile } from "../lockfile";
