+++ src/lockfile/nameAtVersion.ts
@@ -1,0 +1,3 @@
+export function nameAtVersion(name: string, version: string): string {
+  return `${name}@${version}`;
+}
